re-test and check aia-refector

+++ src/modules/attendances/attendances.controller.ts
@@ -1,40 +1,60 @@
-import { Controller, Get, ParseIntPipe, Post } from '@nestjs/common';
+import { BadRequestException, Controller, Get, ParseIntPipe, Post, UseGuards } from '@nestjs/common';
 import { AttendancesService } from './attendances.service';
 import { AttendanceEntity as Attendance } from './attendance.entity';
 import { ApiCreatedResponse, ApiQuery, ApiTags } from '@nestjs/swagger';
 import { Body, Delete, Param, Query, UseFilters } from '@nestjs/common/decorators';
 import { AttendanceDto } from './dto/attendance.dto';
 import { HttpExceptionFilter } from 'src/model/http-exception.filter';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { RolesGuard } from 'src/auth/role.guard';
+import { UserRole } from '@prisma/client';
+import { Roles } from 'src/auth/roles.decorateor';
 
 @Controller('attendances')
 @UseFilters(HttpExceptionFilter)
 @ApiTags('attendances')
 export class AttendancesController {
-  constructor(private readonly attendanceService: AttendancesService) {}
+  constructor(private readonly attendanceService: AttendancesService) { }
+
+
+  @Post()
+  @UseGuards(AuthGuard, RolesGuard)
+  @Roles(UserRole.ADMIN)
+  create(@Body() attendanceDto: AttendanceDto) {
+    const patternTimeFormat = /^([01]\d|2[0-3]):[0-5]\d$/;
+    if (!patternTimeFormat.test(attendanceDto.time)) {
+      throw new BadRequestException('Time not in the correct 24-hour format.');
+    }
+    return this.attendanceService.create(attendanceDto);
+  }
 
   @Get()
   @ApiQuery({ name: 'date', required: false })
-  @ApiQuery({ name: 'id', required: false })
+  @ApiQuery({ name: 'userId', required: false })
   @ApiQuery({ name: 'level', required: false })
-  findAttendances(
+  @UseGuards(AuthGuard, RolesGuard)
+  @Roles(UserRole.SUPER_ADMIN, UserRole.ADMIN)
+  async findAttendances(
     @Query('date') date?: string,
-    @Query('id') id?: string,
+    @Query('userId') userId?: string,
     @Query('level') level?: string,
   ) {
     if (date && level) {
-      return this.attendanceService.findAllByLevelAndDate(level, date);
+      return await this.attendanceService.findAllByLevelAndDate(level, date);
     } else if (date) {
-      return this.attendanceService.findAllByDate(date);
+      return await this.attendanceService.findAllByDate(date);
     } else if (level) {
-      return this.attendanceService.findAllByLevel(level);
-    } else if (id) {
-      return this.attendanceService.findAllByUserId(id);
+      return await this.attendanceService.findAllByLevel(level);
+    } else if (userId) {
+      return await this.attendanceService.findAllByUserId(userId);
     } else {
-      return this.attendanceService.findAll();
+      return await this.attendanceService.findAll();
     }
   }
 
   @Get('calculateAttendance/:date/:userId')
+  @UseGuards(AuthGuard, RolesGuard)
+  @Roles(UserRole.SUPER_ADMIN, UserRole.ADMIN)
   @ApiCreatedResponse({ type: Attendance })
   async calculateAttendance(
     @Param('date') date: string,
@@ -43,13 +63,11 @@ export class AttendancesController {
     return await this.attendanceService.calculateAttendance(date, userId);
   }
 
-  @Post()
-  create(@Body() attendanceDto: AttendanceDto) {
-    return this.attendanceService.create(attendanceDto);
-  }
 
   @Delete(':id')
-  deleteOne(@Param('id', ParseIntPipe) id: number) {
+  async deleteOne(@Param('id', ParseIntPipe) id: number) {
+    const existingAtt = await this.attendanceService.findOneById(id);
+    if (!existingAtt) throw new BadRequestException('Attendance not found');
     return this.attendanceService.deleteOne(id);
   }
 }
